fw/astep24-3l/verification/requirements.txt
cocotbext-uart
pytest
bitstring
PyYAML
Deprecated
